requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
redis[hiredis]>=5.0.0
asyncpg>=0.30.0
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
Pillow>=10.0.0
faiss-cpu>=1.7.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain>=0.3.0
PyPDF2>=3.0.0
pytz>=2024.0
sentence-transformers>=3.0.0
